Move triton kernel benchmarks to tools/benchmark/triton_kernels
Namespace the suite under tools.benchmark.triton_kernels so future
benchmark suites (for non-Triton subsystems) can sit alongside it under
tools/benchmark/. All bench_*.py, runner.py, utils.py, gpu_specs.py, and
__main__.py move into the new subpackage; tools/benchmark/__init__.py
remains as the umbrella package marker.

CLI entry point becomes `python -m tools.benchmark.triton_kernels`.
